Extract 2 party SPDZ setup from test, so that it can be reused later.
