+++ src/test/java/dk/alexandra/fresco/suite/spdz/TestSpdzBasicArithmetic2Parties.java
@@ -32,6 +32,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.logging.Level;
 
+import dk.alexandra.fresco.lib.arithmetic.DivisionTests;
 import org.junit.Test;
 
 import dk.alexandra.fresco.framework.ProtocolEvaluator;
@@ -57,67 +58,8 @@ import dk.alexandra.fresco.suite.spdz.evaluation.strategy.SpdzProtocolSuite;
  * to handle a fixed number of parties.
  * 
  */
-public class TestSpdzBasicArithmetic2Parties {
+public class TestSpdzBasicArithmetic2Parties extends TestSpdz2Parties {
 
-	private static final int noOfParties = 2;
-
-	private void runTest(TestThreadFactory f, EvaluationStrategy evalStrategy,
-			StorageStrategy storageStrategy) throws Exception {
-		Level logLevel = Level.FINE;
-		Reporter.init(logLevel);
-
-		// Since SCAPI currently does not work with ports > 9999 we use fixed
-		// ports
-		// here instead of relying on ephemeral ports which are often > 9999.
-		List<Integer> ports = new ArrayList<Integer>(noOfParties);
-		for (int i = 1; i <= noOfParties; i++) {
-			ports.add(9000 + i);
-		}
-
-		Map<Integer, NetworkConfiguration> netConf = TestConfiguration
-				.getNetworkConfigurations(noOfParties, ports, logLevel);
-		Map<Integer, TestThreadConfiguration> conf = new HashMap<Integer, TestThreadConfiguration>();
-		for (int playerId : netConf.keySet()) {
-			TestThreadConfiguration ttc = new TestThreadConfiguration();
-			ttc.netConf = netConf.get(playerId);
-			
-			SpdzConfiguration spdzConf = new SpdzConfiguration() {
-				
-				@Override
-				public PreprocessingStrategy getPreprocessingStrategy() {
-					return PreprocessingStrategy.DUMMY;
-				}
-
-				@Override
-				public String fuelStationBaseUrl() {
-					return null;
-				}
-				
-				@Override
-				public int getMaxBitLength() {
-					return 150;
-				}
-			};
-			ttc.protocolSuiteConf = spdzConf;
-			boolean useSecureConnection = false; // No tests of secure
-													// connection
-													// here.
-			int noOfVMThreads = 3;
-			int noOfThreads = 3;
-			ProtocolSuite suite = new SpdzProtocolSuite();
-			ProtocolEvaluator evaluator = EvaluationStrategy
-					.fromEnum(evalStrategy);
-			dk.alexandra.fresco.framework.sce.resources.storage.Storage storage = inMemStore;
-			ttc.sceConf = new TestSCEConfiguration(suite, evaluator,
-					noOfThreads, noOfVMThreads, ttc.netConf, storage,
-					useSecureConnection);
-			conf.put(playerId, ttc);
-		}
-		TestThreadRunner.run(f, conf);
-	}
-
-	private static InMemoryStorage inMemStore = new InMemoryStorage();
-	
 	@Test
 	public void test_Copy_Sequential() throws Exception {
 		runTest(new BasicArithmeticTests.TestCopyProtocol(),
